+++ src/main/java/com/incedo/capstone/smartinventory/dto/ReturnsDTO.java
@@ -7,7 +7,7 @@ public class ReturnsDTO {
     private Double quantity;
     private String DateOfDelivery;
     private String DateOfReturn;
-    private int ReceiptNo;
+    private long ReceiptNo;
     private String ReceivedBy;
     private String BillValue;
     private String BillCheckedBy;
@@ -52,11 +52,11 @@ public class ReturnsDTO {
         DateOfReturn = dateOfReturn;
     }
 
-    public int getReceiptNo() {
+    public long getReceiptNo() {
         return ReceiptNo;
     }
 
-    public void setReceiptNo(int receiptNo) {
+    public void setReceiptNo(long receiptNo) {
         ReceiptNo = receiptNo;
     }
